pyproject.toml
[project]
name = "llm-agentic-multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "autogen-agentchat>=0.7.4",
    "autogen-ext[ollama]>=0.7.4",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "llama-index>=0.12.42",
    "llama-index-core>=0.12.42",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-llms-ollama>=0.6.2",
    "llama-parse>=0.6.62",
    "pydantic==2.10.6",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
